An actual test with the PADS layout tool found a subtle error.
Literally the netlist is written ankle bone connects to the leg bone. The leg bone
connects to the thigh bone. The thigh bone ...
You have to repeat the last connection on the next line.
I also delete single pin nets (with message) because they would be
impossible with this system. There is nothing to repeat.